requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[argon2]
python-dotenv
email-validator